refactor(inline): migrate to ds-architecture archetype (SFC split with useInline)
