+++ components/inline/index.ts
@@ -1,4 +1,4 @@
-import Inline from './src/inline.vue';
+import Inline from './src/Inline/index.vue';
 import { withInstall } from '@flash-global66/g-utils';
 import type { SFCWithInstall } from '@flash-global66/g-utils';
 export const GInline: SFCWithInstall<typeof Inline> & {
@@ -8,7 +8,7 @@ export const GInline: SFCWithInstall<typeof Inline> & {
 });
 export default GInline;
 
-export * from './src/inline.type';
-export * from './src/inline';
+export * from './src/types/inline.type';
+export * from './src/Inline/defaults';
 
 export type InlineInstance = InstanceType<typeof Inline>;
