+++ components/inline/package.json
@@ -10,7 +10,7 @@
       "require": "./dist/index.js",
       "types": "./dist/types/index.d.ts"
     },
-    "./styles.scss": "./src/inline.styles.scss",
+    "./styles.scss": "./src/Inline/inline.style.scss",
     "./*": "./*"
   },
   "files": [
