+++ components/inline/src/types/inline.type.ts
